feat: script to automate wordpress installation
Add script to automate the install of the dependencies necessaries for
the installation of wordpress and automate all the process to create the
database in MYSQL that wordpress will be use and the installation of
wordpress perse. The script create the virtual host necessary to create
the site for wordpress with apache2.